src/datapipeline/pyproject.toml
[project]
name = "accimap-data"
version = "0.1.0"
description = "Data pipeline for AcciMap."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage>=3.4.0",
    "pandas>=2.3.3",
    "pyarrow>=21.0.0",
]
